# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /auth/google/callback:
        get:
            tags:
                - AuthService
            operationId: AuthService_Callback
            parameters:
                - name: state
                  in: query
                  schema:
                    type: string
                - name: code
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.CallbackReply'
    /auth/google/login:
        get:
            tags:
                - AuthService
            operationId: AuthService_Auth
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.AuthReply'
    /calendar:
        get:
            tags:
                - CalendarService
            operationId: CalendarService_CreateCalendar
            parameters:
                - name: userID
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.calendar.v1.CreateCalendarReply'
    /login:
        get:
            tags:
                - AuthService
            operationId: AuthService_Login
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.auth.v1.LoginReply'
    /user:
        get:
            tags:
                - UserService
            operationId: UserService_CreateUser
            parameters:
                - name: code
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.CreateUserReply'
components:
    schemas:
        api.auth.v1.AuthReply:
            type: object
            properties:
                url:
                    type: string
        api.auth.v1.CallbackReply:
            type: object
            properties:
                code:
                    type: string
        api.auth.v1.LoginReply:
            type: object
            properties:
                loginPage:
                    type: string
        api.calendar.v1.CreateCalendarReply:
            type: object
            properties:
                calendarID:
                    type: string
        api.user.v1.CreateUserReply:
            type: object
            properties:
                userID:
                    type: string
tags:
    - name: AuthService
    - name: CalendarService
    - name: UserService